implement DMX connection class
Right now the project is dependent on a third-party DMX package that
requires some patching to be usable. The version on PyPI of that DMX
package does not work on Windows. Because that DMX package does not
seem to be maintained any longer (last commit > 2 years ago), the
source on GitHub does not seem to match the PyPI version and the
fact that it is not that hard to implement, this patch implements a
DMX connection class that deprecates the third-party DMX package.

The main advantage of implementing this class ourselves is that
installation will no longer require manual patching of the third-
party DMX package, making installation easier. Secondly, by using
thread-safe data structures we make the sending of packages
somewhat more consistent, hopefully resolving some of the sudden
sudden but short blackouts that occasionally happen now.