Add useful data classes to the package
In this package some specific instances of python's data types are
used. For example, a `bytes()` of exactly three bytes long is used
to represent a color in RGB. Similarly, a `Scene` is a `bytearray()'
of exactly 512 bytes and a `Chase` is a generator that yields
`Scenes`.

Methods and functions in this package are designed with these
particular constraints in mind. This means that there has to
be some sort of mechanism that checks that the data conforms
to these constraints.

Enter the data classes introduced in this patch. These data classes
check for these constraints, so the rest of the functions and
methods in this package do not have to.