Add loopback to DMXConnection and alter termination functionality
To make testing easier, this patch adds loopback functionality to
DMXConnection. By using port 'loop://', a dummified Serial instance
is used. This makes it so the USB-to-DMX hardware does not have
to be plugged in to start an instance of DMXConnection. This makes
testing other parts of this program, like the UI or the chase
generation easier.

Secondly, this patch slightly alters the thread exiting mechanism.
Instead of calling DMXConnection.exit(), the calling thread has to
set the DMXConnection.terminated event using set(), so:
DMXConnection.terminated.set().