Implement ChaseGenerator class
To make extending the package with more chases in the future easier,
it is helpful if all chases are defined in a single location. This
patch provides such a location in the form of the ChaseGenerator
class. It implements all current chase generators as well as a new
one: gen_iso.

An alternate option would have been some sort of configuration file
that defines the chases. However, some chase generators (like
gen_random and gen_iso) are dynamic, and therefore require some
logic to function. By keeping the configuration as code, the full
expressive power of python can be used to define new chases, allowing
for more and more interesting options.