Implement LightControlCLI class
This class implements a replacement for the current CLI. It is less
tightly coupled with the rest of the package, allowing for replacement
by a different sort of UI in the future. Secondly, the help pages
for all commands have been significantly improved. Each command has
its own help entry now, and the CLI will mention when documentation
for a command is missing. Lastly, the CLI uses the `cmd` package
from python's standard library. Using this package gives the advantage
of clear documentation and better comprehensability of this part
of the package.