Remove seemingly unneccesary mark-after-break byte
After testing with a live system, it became apparent that the mark-
after-break byte was causing an off-by-one error. Lamps had to be
set to 2 and 10 instead of 1 and 9. I therefore assume that the
mark-after-byte is sent automatically by pyserial or is being
ignored by the lamps. For this reason, this patch removes the mark-
after-break byte.